Wait for the router and registry pods to come up
The end to end openstack CI started failing intermittently, because
the openshift-ansible finished without waiting for the router to come
up. If we waited with the validation, it would succeed.

So this adds two checks that verify that the registry and router are
actually running or failed before considering the installation done.